tls: Add alt name ostream operators
For printing. And consequential parsing/matching.

Note: because naming of alternative names is inconsistent between tools,
and because openssl is probably more popular when creating certs anyway,
this routine will be inconsistent with both gnutls and openssl (though more
in line with the latter) and name the constants as follows:

* dnsname: "DNS"
* rfc822name: "EMAIL"
* uri: "URI"
* ipaddress "IP"
* othername: "OTHERNAME"
* dn: "DIRNAME"

(cherry picked from commit cdaa9f283202a9591052f39ae8ca9ff6bea653c4)