Do not use std::function for check_for_work()
The check_for_work() function is called in the core reactor loop (and
not just in the path where we are about to enter idle) and is assigned
to a std::function object outside the loop. This forces all invocations
of this function to go through the std::function machinery, which
means that among other things they cannot be inlined.

This does not appear to be intentional, rather the intent seems to have
been to DRY this small function as it is called from two places in
the reactor loop. Instead, use auto type to infer the lambda type and
avoid assignment to std::function, leading to something akin to a direct
call.