tls: Optionally skip client EOF wait
This commit extends the `tls::wrap_client` interface with an optional
config param which may indicates to the session that it should not wait
for an EOF after sending the BYE message.

Badly behaved servers will not send the expected EOF message. Clients
that call `wait_input_shutdown` on the connected socket will hang for 10
seconds in that case.

The previous default behaviour has been preserved and clients should
always wait for EOF if the intend to re-use the connection.