net/tls: Introduce cert_info and accessors
This commit introduces `struct cert_info` which encapsulates useful
information about TLS certificates, currently:

- The certificate's serial number (as assigned by the signing CA)
- The certificate's expiration time

Additionally, this commit adds public API on `tls::certificate_credentials`
to access this information for loaded certificates AND the loaded trust
list.

The motivation for this change is to give applications the ability to
monitor and report on certificate and CA lifetimes on demand, the idea
being that an application may cache serial numbers and expiries when
credentials are reloaded, serving up a summary of those details until
the next reload time.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>