io-queue: Add metrics count for token bucket throttling
Adds a metric counter that counts how many times the io-queue was
throttled from dispatching more events because we failed to grab tokens
from the token bucket.

For cases where requests in the queue and/or average time in the queue
is high this gives a clear signal that that is caused by the tokenbucket
throttling in the io-queue. Hence it allows differentiating from other
cases such as reduced/slow polling.