io-queue: Split throttling metric into feedback vs rate based
Enhances the tokenbucket throttling metric to differentiate between the
two different reasons why the tokenbucket might be throttling.

When getting throttled when grabing tokens from the tokenbucket we look
at the ceil (tracking disk feedback) and head rover (tracking refill
rate). If the difference between ceil and head rover is large enough to
accomodate the pending capacity we count it as rate throttled and disk
feedback throttled otherwise.