net: Fix broken_promise exception on shutdown
- A broken_promise exception was observed when calling close() due to
the fact that the input stream was called while it was waiting on a
pending operation to complete.

- This occurred on call to close because of how do_shutdown() is
implemented. do_shutdown() aquires the output semaphore, but not the
input, then attempts to wait_for_input. If there is already a waiter on
the read path, then the broken_promise exception will be observed when
do_shutdown() calls do_get().

- The fix is to simply modify do_shutdown() to call wait_for_eof()
instead of wait_for_input(),however this would be incomplete because
there are other instances in the implementation where out of band
operations on the read / write path come from opposite paths.

- To resolve the above concern new methods have been
developed (maybe_wait_for_data and maybe_send_data) which attempt to
acquire the in/out semaphore before performing the desired operation. If
the semaphore is already acquired, then the method just returns to the
caller, which would lead to a retry until the operation can be performed.

(cherry picked from commit b754bfaf73ce7f497e3491d66dbbc33853a4574a)