net: Added new functions to support OpenSSL
GnuTLS allows for setting priority and server precedence all
through the GnuTLS priority system.  OpenSSL on the other hand,
has different APIs to handle this.

- To set server precedence in GnuTLS, simply provide the string
  %SERVER_PRECEDENCE in the priority string.  For OpenSSL, the
  flag SSL_OP_CIPHER_SERVER_PREFERENCE must be passed to
  SSL_CTX_set_options

- GnuTLS's priority string can be used to set the priority for
  all versions of TLS.  For OpenSSL, to set the ciphers available
  in TLSv1.2 and below, the user must call
  SSL_CTX_set_cipher_list.
  For TLSv1.3, the call is SSL_CTX_set_ciphersuites

As the format of the cipher/priority strings are different between
OpenSSL and GnuTLS, the user of Seastar would have to change their
code in order to change the format to match the requirements of the
underlying TLS provider.  Therefor, this change adds three new
functions to credentials_builder and server_credentials that plumb
through these strings/settings to OpenSSL and are no-ops in GnuTLS.
Conversely, the set_priority_string function is a no-op when using
OpenSSL.

Signed-off-by: Michael Boquard <michael@redpanda.com>