tls/ossl: Change handle_output_error to not use pointer
While testing OpenSSL in Seastar, it was noticed that certain error
conditions were not being properly handled by Redpanda.  This was due to
how handle_output_error was written to accept `std::exception_ptr`.
When the pointer was re-thrown with `std::throw_with_nested`, the
resulting `std::nested` exception would contain the type
`std::exception_ptr` rather than the actual type of the exception.  This
caused Redpanda's exception handling code to not parse the error
properly.  Because `handle_output_error` could be passed any type of
exception, the function was modified to take in a templatized variable.

Signed-off-by: Michael Boquard <michael@redpanda.com>