net: avoid propagating system errors to errno
When there is an error while trying to write to the underlying socked
inside the bio write method, we should signal that error by propagating
it on openssl's internal error stack rather than through setting errno.

Using the error stack is openssl's conventional way of communicating
errors, so we should use it instead of relying on `errno`. Writing to
`errno` is problematic because it may be read by other parts of the
system and this "poisoning" of the global errno variable could cause
unexpected errors inside other parts of the system.

(cherry picked from commit 526bf5ca2900fa87efb15fa0ec226dc8146fb969)