net/tls: OpenSSL error utilities
Created a new make_ossl_error function that takes in a pre-generated
list of OpenSSL error codes.  This is helpful if the error codes were
already pulled from the stack to be analyzed.

Created a helper function that drains the openssl error stack and
returns a list of errors.

Created a contains_ossl_error function that is used to analyze a list of
errors and determine if any match the library and error code.

Signed-off-by: Michael Boquard <michael@redpanda.com>
(cherry picked from commit d257cfd7853277d1b26645b3e04ed52322b6d339)