optimize abort_source with default exn
When calling abort_source::request_abort() with a default exception, we
can avoid the overhead of creating a new exception object. This is
especially useful in performance-critical code where abort_source is
used frequently.

To achieve this, we modify the request_abort() method to use a static
exception object when the default exception is requested. This way,
we can reuse the same exception object for all calls to request_abort().