stall_detector: no backtrace if exception
If a an exception is in progress when a reactor stall is detected,
omit taking the backtrace, as this can crash as detailed in #2697.

Add a test which reproduces the issue, and which crashes (sometimes)
before this fix and which runs cleanly afterwards.

Fixes #2697.

(cherry picked from commit 7f2aede81b4fa55334a7efdeda3ac5ff58358ea5)