`http`: add `xml` support to `http::mime_types::mappings`
It seems strange that `xml` isn't supported as a `mime` type here.
We expect all S3 REST error responses to be returned with a `Content-Type`
header of `application/xml`, so being able to map the extension here
makes life a lot easier for testing with imposter `http` S3 servers/services.