addr2line: don't load entire file
We support 4 types of input in seastar-addr2line:

1. A list of address, e.g. `0x19c01681 ...` passed on the command line.
2. An explicit file, passed with -f.
3. stdin where stdin is a tty
4. stdin where stdin is not a tty

For all of these cases we load the entire input as a `list` of lines
before starting resolution. This is not necessary and results in a very
large memory requirement, roughly equal to the size of the input file.

It is not unusual to parse 10GB files, which requires more than 10GB
of memory. Instead, simply read the input generator for the types above
line by line. For the specific case of (3) we do still read the entire
input, just to avoid having the output interspersed with input in
the terminal, as we expect it to be small.

Testing on a 1.6 GB file, the RSS is reduced from 2.1 GB to 100 MB, and
runtime is unchanged (tested on cases 2 and 3).