backtrace: move guarded_backtrace, build fix
guarded_backtrace has a static local mutex but was in an anonymous
namespace, which is unsafe as each translation unit would have its own
instance.

Move it to the .cc file (it's happier there anyway) and remove the
anonymous namespace.

Also fix the alpine build by only building this is execinfo is
available.