tls: Catch exception when getting address
There exists a remote possibility that before the client connection is
wrapped in a TLS session, the remote endpoint has closed the connection.
When this occurs, the call to `getpeername` will fail.  The
`connected_socket::remote_address()` method was marked `noexcept` which
caused the program to catch.  The previous commit removed the `noexcept`
decorator and this commit now catches the exception and returns a
default value for the address.

Signed-off-by: Michael Boquard <michael@redpanda.com>